Add get_history_const() function
This returns an immutable reference to the history object. Callers may
need only read access to the history, for example when getting an entry
by index. While callers could use get_history(), which returns a mutable
reference, this causes all callers to get a mutable reference to Editor,
which they may or may not have.