Add AcceptOrNewLine command
This command might be used to insert a newline by default if cursor is
not on the end of line.

This is useful for where multiple lines are not occassional but are used
most of the time (i.e. SQL, Lisp...). The `AcceptLine` only inserts
newline when the input is non-valid, so inserting a thing into the fully
valid chunk of text is cumbersome.