Make custom key bindings an optional Cargo feature
Add a cargo feature called `custom-bindings` which is enabled by
default. The `custom-bindings` feature enables APIs which allow users of
`rustyline` to define custom keybindings.

Custom keybindings are the only use of `radix_trie` in this crate. If
the `custom-bindings` feature is disabled, `radix_tree` is not included
in `rustyline`.

Motivation of this commit is to reduce dependencies.

The approach I've taken in this commit is to stub out internal APIs when
`custom-bindings` feature is not active to minimize the number of code
changes that need to occur. Similarly, when the `custom-bindings`
feature is not enabled, I've added `allow(dead_code)` to `mod bindings`.

The splitting out of default functionality into a new feature that is
enabled by default I believe is a breaking change, because users may
have previously been using custom keybindings with default-features =
false.

This commit adds a step to the GitHub Actions CI to run `cargo check`
with no default features.