Use lookup tables instead of linear search with `memchr`
Several places use `memchr` with a byteset. This commit refactors these
code paths to construct a lookup table from `u8` -> `bool`, where an
index is set to `true` if the byte is present in the given slice.

This change removes linear scans that occur in loops, which changes
these functions runtime complexity from `O(m * n)` to `O(m + n)`.