Fix position calculation on Windows
The implementations for Renderer::calculate_position differed slightly
between unix and Windows. Notably, the Windows implementation did not
use tty::width, but instead used UnicodeWidthStr::width. It seems that
the latter did not correctly compute the length.

For context, this bug was discovered in the boa JS interpreter, where on
Windows the prompt being bold and blue caused the cursor to be offset by
9 characters.